+++ chain/tests/mine_simple_chain.rs
@@ -32,7 +32,7 @@ use core::core::target::Difficulty;
 use core::core::{Block, BlockHeader, OutputFeatures, OutputIdentifier, Transaction};
 use core::global;
 use core::global::ChainTypes;
-use wallet::libtransaction::{self, build};
+use wallet::libtx::{self, build};
 
 use keychain::Keychain;
 
@@ -64,7 +64,7 @@ fn mine_empty_chain() {
 		let prev = chain.head_header().unwrap();
 		let difficulty = consensus::next_difficulty(chain.difficulty_iter()).unwrap();
 		let pk = keychain.derive_key_id(n as u32).unwrap();
-		let reward = libtransaction::reward::output(&keychain, &pk, 0, prev.height).unwrap();
+		let reward = libtx::reward::output(&keychain, &pk, 0, prev.height).unwrap();
 		let mut b = core::core::Block::new(&prev, vec![], difficulty.clone(), reward).unwrap();
 		b.header.timestamp = prev.timestamp + time::Duration::seconds(60);
 
@@ -413,7 +413,7 @@ fn prepare_block_nosum(
 	let key_id = kc.derive_key_id(diff as u32).unwrap();
 
 	let fees = txs.iter().map(|tx| tx.fee()).sum();
-	let reward = libtransaction::reward::output(&kc, &key_id, fees, prev.height).unwrap();
+	let reward = libtx::reward::output(&kc, &key_id, fees, prev.height).unwrap();
 	let mut b = match core::core::Block::new(prev, txs, Difficulty::from_num(diff), reward) {
 		Err(e) => panic!("{:?}", e),
 		Ok(b) => b,
