+++ core/tests/common/mod.rs
@@ -23,8 +23,8 @@ use grin_core::core::Transaction;
 use grin_core::core::block::{Block, BlockHeader};
 use grin_core::core::target::Difficulty;
 use keychain::{Identifier, Keychain};
-use wallet::libtransaction::build::{self, input, output, with_fee};
-use wallet::libtransaction::reward;
+use wallet::libtx::build::{self, input, output, with_fee};
+use wallet::libtx::reward;
 
 // utility producing a transaction with 2 inputs and a single outputs
 pub fn tx2i1o() -> Transaction {
