+++ pool/tests/graph.rs
@@ -24,7 +24,7 @@ extern crate rand;
 use core::core::OutputFeatures;
 use core::core::transaction::ProofMessageElements;
 use keychain::Keychain;
-use wallet::libtransaction::proof;
+use wallet::libtx::proof;
 
 #[test]
 fn test_add_entry() {
