+++ wallet/src/libwallet/selection.rs
@@ -16,7 +16,7 @@
 
 use failure::ResultExt;
 use keychain::{Identifier, Keychain};
-use libtransaction::{build, tx_fee, slate::Slate};
+use libtx::{build, tx_fee, slate::Slate};
 use libwallet::types::*;
 use libwallet::{keys, sigcontext};
 use types::*;
