+++ wallet/src/libwallet/sigcontext.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 //! Signature context holder helper (may be removed or replaced eventually)
 use keychain::extkey::Identifier;
-use libtransaction::aggsig;
+use libtx::aggsig;
 use util::secp::key::{PublicKey, SecretKey};
 use util::secp::{self, Secp256k1};
